+++ package.json
@@ -29,7 +29,7 @@
     "tsc": "tsc",
     "tsc-watch": "tsc -w",
     "collab": "cross-env HOST=localhost PORT=1234 npx y-websocket-server",
-    "validation": "npx tsc packages/lexical-playground/src/server/validation.ts",
+    "validation": "node packages/lexical-playground/src/server/validation.js",
     "test-unit": "jest --selectProjects unit --testPathPattern",
     "test-unit-watch": "npm run test-unit -- --watch --coverage=''",
     "test-e2e-chromium": "cross-env E2E_BROWSER=chromium playwright test --project=\"chromium\"",
